Rename RenderLoop to AppLoop
